Add: reservation filter style query in welcome
